requirements.txt
streamlit
openai==1.43.0
pandas
faker
python-dotenv
transformers
chromadb
sentence-transformers
deepeval
